requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
ply==3.11
pyasn1==0.6.0
pycparser==2.22
pysmi-lextudio==1.4.3
pysnmp-lextudio==6.1.2
pysnmpcrypto==0.0.4
requests==2.31.0
urllib3==2.2.1
